styling: fix wrap for plan details
